requirements.txt
docker-py
docopt
dockerpty
Jinja2